Refactor ProposalBuilder Component and Update Imports for Proposal Creation Pages
- Removed the ProposalBuilder component from its original location, streamlining the import paths across multiple proposal creation pages.
- Updated the import statements in SafeCreateProposalTemplatePage, SafeProposalWithActionsCreatePage, SafeProposalCreatePage, and SafeSablierProposalCreatePage to reflect the new structure.
- Enhanced code organization and maintainability by consolidating related components under a single directory.

These changes aim to improve the modularity and clarity of the proposal creation process, facilitating easier future enhancements.
